+++ tests/PandoTests/Tests/Repositories/InMemoryRepositoryTests.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
 using Pando;
+using Pando.DataSources;
+using Pando.DataSources.Utils;
 using Pando.Exceptions;
-using Pando.Repositories;
-using Pando.Repositories.Utils;
 using PandoTests.Utils;
 using Standart.Hash.xxHash;
 using Xunit;
