+++ tests/PandoTests/Tests/Repositories/StreamRepositoryTests.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
-using Pando.Repositories;
-using Pando.Repositories.Utils;
+using Pando.DataSources;
+using Pando.DataSources.Utils;
 using PandoTests.Utils;
 using Standart.Hash.xxHash;
 using Xunit;
